// src/service/modules/assessmentTask/type.ts
interface BoundSubProcess {
  processId: number
  subProcessCode: number
  evaluationTarget: number
  dataSourceCol: string
  dataSourceId: number
  inputParamCol: string
  defaultValue: string
  evaluationTaskId: string
  filterCriteria: string
}